Use better factorial approximation
Previously, we simply used the Stirling Approximation. Now, we reduce the input by recurrence until it is in the interval [0, 1]. Then, we use a Cubic Approximation to calculate the result. The Cubic Approximation is constructed to have f(0) = 1 and f(1) = 1. Furthermore, for p = (p_x, p_y) the minimum point of the factorial on the interval [0, 1], it is constructed to have f(p_x) = p_y and f'(p_x) = 0.